requirements.txt
fastapi
numpy
tensorflow
pillow